+++ methods/src/test/kotlin/methods/equations/LogarithmicEquationsTest.kt
@@ -22,6 +22,7 @@ import methods.algebra.AlgebraExplanation
 import methods.logs.LogsExplanation
 import kotlin.test.Test
 
+@Suppress("LargeClass")
 class LogarithmicEquationsTest {
     // Right now these are only for the normalization part, but should be updated once
     // we have solutions implemented.
@@ -80,7 +81,7 @@ class LogarithmicEquationsTest {
             check {
                 task {
                     explanation {
-                        key = AlgebraExplanation.ComputeDomainOfAlgebraicExpression
+                        key = AlgebraExplanation.ComputeDomainOfEquation
                     }
 
                     step {
@@ -130,7 +131,7 @@ class LogarithmicEquationsTest {
             check {
                 task {
                     explanation {
-                        key = AlgebraExplanation.ComputeDomainOfAlgebraicExpression
+                        key = AlgebraExplanation.ComputeDomainOfEquation
                     }
 
                     step {
@@ -169,6 +170,70 @@ class LogarithmicEquationsTest {
 
                 task {
                     startExpr = "SetSolution[x: {[/e/ ^ 3] + 2}] GIVEN SetSolution[x: (2, /infinity/)]"
+                }
+            }
+        }
+
+        testMethodInX {
+            method = EquationsPlans.SolveEquation
+            inputExpr = "log (x + 7) - 5 = 0"
+
+            check {
+                task {
+                    explanation {
+                        key = AlgebraExplanation.ComputeDomainOfEquation
+                    }
+
+                    step {
+                        toExpr = "SetSolution[x: (-7, /infinity/)]"
+                    }
+                }
+
+                task {
+                    step {
+                        explanation {
+                            key = EquationsExplanation.SolveLogarithmicEquations
+                        }
+
+                        step {
+                            toExpr = "log (x + 7) = 5"
+                        }
+
+                        step {
+                            toExpr = "[10 ^ log (x + 7)] = [10 ^ 5]"
+                            explanation {
+                                key = LogsExplanation.ExponentiateBothSides
+                            }
+                        }
+
+                        step {
+                            toExpr = "x + 7 = [10 ^ 5]"
+                            explanation {
+                                key = LogsExplanation.SimplifyLogInExponentWithMatchingBase
+                            }
+                        }
+
+                        step {
+                            toExpr = "SetSolution[x: {99993}]"
+                            explanation {
+                                key = EquationsExplanation.SolveLinearEquation
+                            }
+                        }
+                    }
+                }
+
+                task {
+                    startExpr = "SetSolution[x: {99993}] GIVEN SetSolution[x: (-7, /infinity/)]"
+                    explanation {
+                        key = EquationsExplanation.AddDomainConstraintToSolution
+                    }
+
+                    step {
+                        toExpr = "SetSolution[x: {99993}]"
+                        explanation {
+                            key = EquationsExplanation.GatherSolutionsAndConstraint
+                        }
+                    }
                 }
             }
         }
@@ -229,6 +294,160 @@ class LogarithmicEquationsTest {
 
                 task {
                     startExpr = "SetSolution[x : {[10 ^ -[3 / 7]] - 1}] GIVEN SetSolution[x: (-1, /infinity/)]"
+                }
+            }
+        }
+    }
+
+    @Suppress("LongMethod")
+    @Test
+    fun `move negative constant terms and express as logs`() {
+        testMethodInX {
+            method = EquationsPlans.SolveEquation
+            inputExpr = "log_[2] x = log_[2] 7 - 2"
+
+            check {
+                task {
+                    explanation {
+                        key = AlgebraExplanation.ComputeDomainOfEquation
+                    }
+
+                    step {
+                        toExpr = "SetSolution[x: (0, /infinity/)]"
+                    }
+                }
+
+                task {
+                    step {
+                        explanation {
+                            key = EquationsExplanation.SolveLogarithmicEquations
+                        }
+
+                        step {
+                            toExpr = "log_[2] x + 2 = log_[2] 7"
+                            explanation {
+                                key = LogsExplanation.MoveNegatedConstantTermsToTheOppositeSideAndSimplify
+                            }
+                        }
+
+                        step {
+                            toExpr = "log_[2] x + log_[2] 4 = log_[2] 7"
+                            explanation {
+                                key = LogsExplanation.ExpressConstantInSumAsLogAndSimplify
+                            }
+                        }
+
+                        step {
+                            toExpr = "log_[2] (4 x) = log_[2] 7"
+                            explanation {
+                                key = LogsExplanation.CollectLogarithmsInSum
+                            }
+                        }
+
+                        step {
+                            toExpr = "4 x = 7"
+                            explanation {
+                                key = LogsExplanation.ApplyEqualityRuleOfLogs
+                            }
+                        }
+
+                        step {
+                            toExpr = "SetSolution[x: {[7 / 4]}]"
+                            explanation {
+                                key = EquationsExplanation.SolveLinearEquation
+                            }
+                        }
+                    }
+                }
+
+                task {
+                    explanation {
+                        key = EquationsExplanation.AddDomainConstraintToSolution
+                    }
+
+                    step {
+                        fromExpr = "SetSolution[x: {[7 / 4]}] GIVEN SetSolution[x: (0, /infinity/)]"
+                        toExpr = "SetSolution[x: {[7 / 4]}]"
+                        explanation {
+                            key = EquationsExplanation.GatherSolutionsAndConstraint
+                        }
+                    }
+                }
+            }
+        }
+
+        testMethodInX {
+            method = EquationsPlans.SolveEquation
+            inputExpr = "log (x + 1) - log (x - 2) - 3 = 0"
+
+            check {
+                task {
+                    explanation {
+                        key = AlgebraExplanation.ComputeDomainOfEquation
+                    }
+
+                    step {
+                        toExpr = "SetSolution[x: (2, /infinity/)]"
+                    }
+                }
+
+                task {
+                    step {
+                        explanation {
+                            key = EquationsExplanation.SolveLogarithmicEquations
+                        }
+
+                        step {
+                            toExpr = "log (x + 1) - 3 = log (x - 2)"
+                            explanation {
+                                key = LogsExplanation.MoveNegatedLogarithmicTermsToOppositeSideAndSimplify
+                            }
+                        }
+
+                        step {
+                            toExpr = "log (x + 1) = log (x - 2) + 3"
+                            explanation {
+                                key = LogsExplanation.MoveNegatedConstantTermsToTheOppositeSideAndSimplify
+                            }
+                        }
+
+                        step {
+                            toExpr = "log (x + 1) = log (x - 2) + log 1000"
+                            explanation {
+                                key = LogsExplanation.ExpressConstantInSumAsLogAndSimplify
+                            }
+                        }
+
+                        step {
+                            toExpr = "log (x + 1) = log (1000 (x - 2))"
+                            explanation {
+                                key = LogsExplanation.CollectLogarithmsInSum
+                            }
+                        }
+
+                        step {
+                            toExpr = "x + 1 = 1000 (x - 2)"
+                            explanation {
+                                key = LogsExplanation.ApplyEqualityRuleOfLogs
+                            }
+                        }
+
+                        step {
+                            toExpr = "SetSolution[x: {[667 / 333]}]"
+                            explanation {
+                                key = EquationsExplanation.SolveLinearEquation
+                            }
+                        }
+                    }
+                }
+
+                task {
+                    step {
+                        toExpr = "SetSolution[x: {[667 / 333]}]"
+                        explanation {
+                            key = EquationsExplanation.GatherSolutionsAndConstraint
+                        }
+                    }
                 }
             }
         }
@@ -333,12 +552,23 @@ class LogarithmicEquationsTest {
                                 key = LogsExplanation.CollectLogarithmsInSum
                             }
                         }
+
+                        step {
+                            toExpr = "2 x * [(x - 2) ^ 2] = 25 * [(x + 1) ^ 6]"
+                            explanation {
+                                key = LogsExplanation.ApplyEqualityRuleOfLogs
+                            }
+                        }
+
+                        step {
+                            toExpr = "2 [x^3] - 8 [x^2] + 8x - 25 * [(x + 1) ^ 6] = 0"
+                        }
                     }
                 }
 
                 task {
                     startExpr =
-                        "log_[2] (2 x * [(x - 2) ^ 2]) = log_[2] (25 * [(x + 1) ^ 6]) " +
+                        "2 [x^3] - 8 [x^2] + 8x - 25 * [(x + 1) ^ 6] = 0 " +
                         "GIVEN SetSolution[x: (2, /infinity/)]"
                 }
             }
@@ -356,24 +586,31 @@ class LogarithmicEquationsTest {
                     step {
                         toExpr = "SetSolution[x: (-1, /infinity/)] AND [x ^ 3] + 3 x + 7 > 0"
                         explanation {
-                            key = AlgebraExplanation.ComputeDomainOfAlgebraicExpression
+                            key = AlgebraExplanation.ComputeDomainOfEquation
                         }
                     }
                 }
 
                 task {
                     step {
-                        fromExpr = "3 * log (x + 1) = log ([x ^ 3] + 3 x + 7)"
-                        toExpr = "log [(x + 1) ^ 3] = log ([x ^ 3] + 3 x + 7)"
-                        explanation {
-                            key = LogsExplanation.RewriteCoefficientsAsExponents
+                        step {
+                            toExpr = "log [(x + 1) ^ 3] = log ([x ^ 3] + 3 x + 7)"
+                            explanation {
+                                key = LogsExplanation.RewriteCoefficientsAsExponents
+                            }
+                        }
+                        step {
+                            toExpr = "[(x + 1) ^ 3] = [x ^ 3] + 3 x + 7"
+                        }
+                        step {
+                            toExpr = "SetSolution[x : {-sqrt[2], sqrt[2]}]"
                         }
                     }
                 }
 
                 task {
                     startExpr =
-                        "log [(x + 1) ^ 3] = log ([x ^ 3] + 3 x + 7) " +
+                        "SetSolution[x : {-sqrt[2], sqrt[2]}] " +
                         "GIVEN SetSolution[x: (-1, /infinity/)] AND [x ^ 3] + 3 x + 7 > 0"
                     explanation {
                         key = EquationsExplanation.AddDomainConstraintToSolution
@@ -392,7 +629,7 @@ class LogarithmicEquationsTest {
             check {
                 task {
                     explanation {
-                        key = AlgebraExplanation.ComputeDomainOfAlgebraicExpression
+                        key = AlgebraExplanation.ComputeDomainOfEquation
                     }
 
                     step {
@@ -402,7 +639,6 @@ class LogarithmicEquationsTest {
 
                 task {
                     step {
-                        toExpr = "log (9 - x) = log (9 x)"
                         explanation {
                             key = EquationsExplanation.SolveLogarithmicEquations
                         }
@@ -417,11 +653,24 @@ class LogarithmicEquationsTest {
                                 key = LogsExplanation.CollectLogarithmsInSum
                             }
                         }
+
+                        step {
+                            toExpr = "9 - x = 9 x"
+                        }
+
+                        step {
+                            toExpr = "SetSolution[x : {[9 / 10]}]"
+                        }
                     }
                 }
 
                 task {
-                    startExpr = "log (9 - x) = log (9 x) GIVEN SetSolution[x: (0, 9)]"
+                    step {
+                        toExpr = "SetSolution[x: {[9 / 10]}]"
+                        explanation {
+                            key = EquationsExplanation.GatherSolutionsAndConstraint
+                        }
+                    }
                 }
             }
         }
@@ -436,7 +685,7 @@ class LogarithmicEquationsTest {
             check {
                 task {
                     explanation {
-                        key = AlgebraExplanation.ComputeDomainOfAlgebraicExpression
+                        key = AlgebraExplanation.ComputeDomainOfEquation
                     }
 
                     step {
@@ -461,11 +710,136 @@ class LogarithmicEquationsTest {
                         step {
                             toExpr = "log (x (x - 16)) = log 105"
                         }
+
+                        step {
+                            toExpr = "x (x - 16) = 105"
+                        }
+
+                        step {
+                            toExpr = "SetSolution[x : {-5, 21}]"
+                        }
                     }
                 }
 
                 task {
-                    startExpr = "log (x (x - 16)) = log 105 GIVEN SetSolution[x: (16, /infinity/)]"
+                    startExpr = "SetSolution[x : {-5, 21}] GIVEN SetSolution[x: (16, /infinity/)]"
+
+                    step {
+                        toExpr = "SetSolution[x: {21}]"
+                        explanation {
+                            key = EquationsExplanation.GatherSolutionsAndConstraint
+                        }
+                    }
+                }
+            }
+        }
+    }
+
+    @Suppress("LongMethod")
+    @Test
+    fun `eliminate invalid solutions in eq with multiple constraints`() {
+        testMethodInX {
+            method = EquationsPlans.SolveEquation
+            inputExpr = "3 * log (x + 1) = log ([x ^ 3] + 3 x + 7)"
+
+            check {
+                task {
+                    explanation {
+                        key = AlgebraExplanation.ComputeDomainOfEquation
+                    }
+
+                    step {
+                        toExpr = "SetSolution[x: (-1, /infinity/)] AND [x ^ 3] + 3 x + 7 > 0"
+                    }
+                }
+
+                task {
+                    step {
+                        explanation {
+                            key = EquationsExplanation.SolveLogarithmicEquations
+                        }
+
+                        step {
+                            toExpr = "log [(x + 1) ^ 3] = log ([x ^ 3] + 3 x + 7)"
+                            explanation {
+                                key = LogsExplanation.RewriteCoefficientsAsExponents
+                            }
+                        }
+
+                        step {
+                            toExpr = "[(x + 1) ^ 3] = [x ^ 3] + 3 x + 7"
+                            explanation {
+                                key = LogsExplanation.ApplyEqualityRuleOfLogs
+                            }
+                        }
+
+                        step {
+                            toExpr = "SetSolution[x: {-sqrt[2], sqrt[2]}]"
+
+                            explanation {
+                                key = EquationsExplanation.SolveEquationUsingRootsMethod
+                            }
+                        }
+                    }
+                }
+
+                task {
+                    taskId = "#3"
+                    startExpr = "SetSolution[x: {-sqrt[2], sqrt[2]}] GIVEN " +
+                        "SetSolution[x: (-1, /infinity/)] AND [x ^ 3] + 3 x + 7 > 0"
+                    explanation {
+                        key = EquationsExplanation.AddDomainConstraintToSolution
+                    }
+
+                    step {
+                        toExpr = "SetSolution[x: {sqrt[2]}]"
+                        explanation {
+                            key = EquationsExplanation.GatherSolutionsAndConstraint
+                        }
+
+                        task {
+                            taskId = "#1"
+                            startExpr = "[(-sqrt[2]) ^ 3] + 3 (-sqrt[2]) + 7 > 0"
+                            explanation {
+                                key = EquationsExplanation.CheckIfSolutionSatisfiesConstraint
+                            }
+
+                            step {
+                                toExpr = "-5 sqrt[2] + 7 > 0"
+                            }
+
+                            step {
+                                toExpr = "-0.0710678119 > 0"
+                            }
+
+                            step {
+                                toExpr = "Contradiction[-0.0710678119 > 0]"
+                            }
+                        }
+
+                        task {
+                            taskId = "#2"
+                            startExpr = "[(sqrt[2]) ^ 3] + 3 sqrt[2] + 7 > 0"
+                            explanation {
+                                key = EquationsExplanation.CheckIfSolutionSatisfiesConstraint
+                            }
+
+                            step {
+                                toExpr = "5 sqrt[2] + 7 > 0"
+                            }
+
+                            step {
+                                toExpr = "Identity[5 sqrt[2] + 7 > 0]"
+                            }
+                        }
+
+                        task {
+                            startExpr = "SetSolution[x: {sqrt[2]}]"
+                            explanation {
+                                key = EquationsExplanation.SomeSolutionsDoNotSatisfyConstraint
+                            }
+                        }
+                    }
                 }
             }
         }