+++ CHANGELOG.md
@@ -13,6 +13,7 @@ contents to a new release section (e.g. _[0.x]_).
 
 - Normalize equations with logarithmic terms. (PLUT-1120)
 - Solve elementary logarithmic equations. (PLUT-1123)
+- Solve logarithmic equations of form log(f) = log(g). (PLUT-1125)
 
 ### Fixed